phone.py
Amidst so much technology, it has become easier to access websites and check product prices. And with Python, it's even better. The code is simple; it accesses four websites and extracts the product name, phone number, and price, stores them in a variable, and sends them to Excel, creating new sheets organized by brand.